REF: organize Series indexing tests (#31557)
