+++ pandas/tests/series/methods/test_convert_dtypes.py
